+++ R/read_softmax5_template.R
@@ -6,8 +6,14 @@ read_softmax5_template <- function(file,
                                    wellsAsFactors = TRUE,
                                    groupsAsFactors = TRUE,
                                    typesAsFactors = TRUE,
-                                   #encoding = "UTF-16LE", #ignored
+                                   encoding = "guess",
                                    ...) {
+
+    if (encoding == "guess") {
+        encoding <- readr::guess_encoding(file)[[1, "encoding"]]
+        message(sprintf("Using encoding '%s'", encoding))
+    }
+
     d <- utils::read.delim(
         file = file,
         header = FALSE,